thank you malum

+++ src/main/java/earth/terrarium/pastel/mixin/SoulWardDataMixin.java
@@ -14,7 +14,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(SoulWardData.class)
 public abstract class SoulWardDataMixin {
     @Shadow
-    public abstract void addCooldown(LivingEntity living, double multiplier);
+    public abstract void addCooldown(LivingEntity living, float multiplier);
 
     @Inject(method = "recoverSoulWard", at = @At("HEAD"), cancellable = true)
     private void soulWardDisruptedByStakes(LivingEntity entity, double amount, CallbackInfo ci) {
